% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{ABLTAG_FIELD_METADATA_TABLE_FIELDS}
\alias{ABLTAG_FIELD_METADATA_TABLE_FIELDS}
\title{Fields describing the metadata for each field uploaded to the DB}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
ABLTAG_FIELD_METADATA_TABLE_FIELDS
}
\description{
Fields describing the metadata for each field uploaded to the DB
}
\keyword{datasets}
