% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{ABLTAG_TABLE_METADATA_TABLE_FIELDS}
\alias{ABLTAG_TABLE_METADATA_TABLE_FIELDS}
\title{Fields for metadata describing the tables to which these data are written}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
ABLTAG_TABLE_METADATA_TABLE_FIELDS
}
\description{
Fields for metadata describing the tables to which these data are written
}
\keyword{datasets}
