% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{AUTOGENERATED_FIELD_METADATA_FIELDS}
\alias{AUTOGENERATED_FIELD_METADATA_FIELDS}
\title{Data automatically generated by the system to describe the fields in the outgoing data}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
AUTOGENERATED_FIELD_METADATA_FIELDS
}
\description{
Data automatically generated by the system to describe the fields in the outgoing data
}
\keyword{datasets}
