% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{AUTOGENERATED_METADATA_FIELDS}
\alias{AUTOGENERATED_METADATA_FIELDS}
\title{Fields automatically provided by the metadata DataMap}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
AUTOGENERATED_METADATA_FIELDS
}
\description{
Fields automatically provided by the metadata DataMap
}
\keyword{datasets}
