% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_field_metadata_frame}
\alias{build_field_metadata_frame}
\title{Build a data.frame of field metadata from a Decoder}
\usage{
build_field_metadata_frame(dc)
}
\arguments{
\item{dc}{The \code{Decoder} to build the metadata frame from}
}
\value{
A data.frame of field metadata
}
\description{
Build a data.frame of field metadata from a Decoder
}
