% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{build_temp_db}
\alias{build_temp_db}
\title{Build a temporary testing DB}
\usage{
build_temp_db(
  from_ = here::here("db", "temp_db_snapshot.db"),
  to_ = withr::local_tempfile(pattern = "metl_test_db"),
  dats = list(),
  ...
)
}
\arguments{
\item{from_}{Path of DB snapshot}

\item{to_}{Path of temporary DB}

\item{dats}{Named list of data.frames to use to populate the DB.}
}
\description{
Builds a temporary DB based on a snapshot. By default DB saved in temporary directory
}
