% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{bump_timestamps}
\alias{bump_timestamps}
\title{Make timestamps unique}
\usage{
bump_timestamps(v, incr = 1e-06)
}
\arguments{
\item{v}{The vector of timestamps (as integers)}
}
\value{
The same vectpr of timestamps as \code{v}, in the same order, but with any identical timestamps having been bumped to unique values.
}
\description{
Helper function for making timestamps unique. This is accomplished by looking for groups of identical timestamps within \code{v}, and if any are found each member of the group is incremented by \code{incr} so as to become a unique value. By default, \code{incr} is 1e-6 s, the smallest possible value which a POSIXct timestamp can record, and so the smallest differential by which two timestamps can be considered to be unique to one another.
}
