% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier.R
\docType{class}
\name{Condition-class}
\alias{Condition-class}
\alias{Condition}
\title{Condition class.}
\description{
Used to define a condition the conditions which a directory must satisfy to be considered a valid tag directory.#'
}
\section{Fields}{

\describe{
\item{\code{condition}}{function. Function which returns TRUE if the directory satisfies the condition, FALSE otherwise.}

\item{\code{message}}{character. Message to be displayed if the condition is not satisfied.}
}}


