% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{DataMap-class}
\alias{DataMap-class}
\alias{DataMap}
\title{DataMap}
\value{
An instance of the DataMap class
}
\description{
Provides base structure for all make/model specific tag data maps
Analagous to a table in the final DB. Desribes programatically how the data
to be inserted into that table should be collected from the disk, and
transformed.
Not intended to be implemented directly
}
\section{Fields}{

\describe{
\item{\code{d}}{character. The directory in which the tag data can be found}

\item{\code{input_data_field_map}}{FieldMap. Map of the data fields and their original format}

\item{\code{output_data_field_map}}{FieldMap. Map of the data fields and their corresponding DB format}
}}

\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

