% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_DesertStar_SeaTagMOD-class}
\alias{DataMap_DesertStar_SeaTagMOD-class}
\alias{DataMap_DesertStar_SeaTagMOD}
\title{DataMap - Desert Star SeaTag MOD data - base class}
\description{
DataMap - Desert Star SeaTag MOD data - base class
}

