% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_DesertStar_SeaTagMOD_InstantSensorData-class}
\alias{DataMap_DesertStar_SeaTagMOD_InstantSensorData-class}
\alias{DataMap_DesertStar_SeaTagMOD_InstantSensorData}
\title{DataMap - Desert Star SeaTag MOD instant sensor data}
\description{
DataMap - Desert Star SeaTag MOD instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

