% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_DesertStar_SeaTagMOD_InstantSensorData}
\alias{DataMap_DesertStar_SeaTagMOD_InstantSensorData}
\title{DataMap - Desert Star SeaTag MOD instant sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_DesertStar_SeaTagMOD_InstantSensorData
}
\description{
DataMap - Desert Star SeaTag MOD instant sensor data
}
\keyword{datasets}
