% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_DesertStar_SeaTagMOD_InstantSensorData_Directory}
\alias{DataMap_DesertStar_SeaTagMOD_InstantSensorData_Directory}
\title{DataMap - Desert Star SeaTag MOD directory extractor}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_DesertStar_SeaTagMOD_InstantSensorData_Directory
}
\description{
DataMap - Desert Star SeaTag MOD directory extractor
}
\keyword{datasets}
