% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_DesertStar_SeaTagMOD_TagMetaData-class}
\alias{DataMap_DesertStar_SeaTagMOD_TagMetaData-class}
\alias{DataMap_DesertStar_SeaTagMOD_TagMetaData}
\title{DataMap - Desert Star SeaTag MOD tag metadata}
\value{
The tag ID identified from the files, as a string
}
\description{
DataMap - Desert Star SeaTag MOD tag metadata
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

