% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_DesertStar_SeaTagMOD_TagMetaData}
\alias{DataMap_DesertStar_SeaTagMOD_TagMetaData}
\title{DataMap - Desert Star SeaTag MOD tag metadata}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_DesertStar_SeaTagMOD_TagMetaData
}
\description{
DataMap - Desert Star SeaTag MOD tag metadata
}
\keyword{datasets}
