% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_DesertStar_TagMetaData-class}
\alias{DataMap_DesertStar_TagMetaData-class}
\alias{DataMap_DesertStar_TagMetaData}
\title{DataMap - Desert Star tag metadata - base class}
\description{
DataMap - Desert Star tag metadata - base class
}

