% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_InstantSensorData_Base-class}
\alias{DataMap_InstantSensorData_Base-class}
\alias{DataMap_InstantSensorData_Base}
\title{DataMap - Instant sensor data - base class}
\description{
DataMap - Instant sensor data - base class
}

