% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1000.1100.1250_InstantSensorData-class}
\alias{DataMap_Lotek.1000.1100.1250_InstantSensorData-class}
\alias{DataMap_Lotek.1000.1100.1250_InstantSensorData}
\title{DataMap for the Lotek 1000/1100/1250 tags}
\description{
DataMap for the Lotek 1000/1100/1250 tags
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}

\item{\code{read_csv_lotek_1000.1100.1250(fp)}}{Read csv data from Lotek 1000/1100/1250 formatted file}
}}

