% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1000.1100.1250_TagMetaData-class}
\alias{DataMap_Lotek.1000.1100.1250_TagMetaData-class}
\alias{DataMap_Lotek.1000.1100.1250_TagMetaData}
\title{DataMap - Lotek 1000/1100/1250 tag metadata}
\description{
DataMap - Lotek 1000/1100/1250 tag metadata
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

