% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1300_InstantSensorData-class}
\alias{DataMap_Lotek.1300_InstantSensorData-class}
\alias{DataMap_Lotek.1300_InstantSensorData}
\title{DataMap - Instant sensor data from Lotek 1300 tags}
\description{
DataMap - Instant sensor data from Lotek 1300 tags
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

