% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1300_TagMetaData-class}
\alias{DataMap_Lotek.1300_TagMetaData-class}
\alias{DataMap_Lotek.1300_TagMetaData}
\title{DataMap - Tag metadata from Lotek 1300 tags}
\description{
DataMap - Tag metadata from Lotek 1300 tags
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

