% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1400.1800_InstantSensorData-class}
\alias{DataMap_Lotek.1400.1800_InstantSensorData-class}
\alias{DataMap_Lotek.1400.1800_InstantSensorData}
\title{Datamap - Lotek 1400/1800 instant sensor data}
\description{
Datamap - Lotek 1400/1800 instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}

\item{\code{read_csv_lotek_1400.1800(fp)}}{Read csv data from Lotek 1400/1800 formatted csv file}
}}

