% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek.1400.1800_TagMetaData-class}
\alias{DataMap_Lotek.1400.1800_TagMetaData-class}
\alias{DataMap_Lotek.1400.1800_TagMetaData}
\title{DataMap - Tag metadata from Lotek 1400/1800 tags}
\description{
DataMap - Tag metadata from Lotek 1400/1800 tags
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

