% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_Lotek_1000.1100.1250_InstantSensorData}
\alias{DataMap_Lotek_1000.1100.1250_InstantSensorData}
\title{DataMap for the Lotek 1000/1100/1250 tags}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_Lotek_1000.1100.1250_InstantSensorData
}
\description{
DataMap for the Lotek 1000/1100/1250 tags
}
\keyword{datasets}
