% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_Lotek_1000.1100.1250_TagMetaData}
\alias{DataMap_Lotek_1000.1100.1250_TagMetaData}
\title{DataMap - Lotek 1000/1100/1250 tag metadata}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_Lotek_1000.1100.1250_TagMetaData
}
\description{
DataMap - Lotek 1000/1100/1250 tag metadata
}
\keyword{datasets}
