% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_Lotek_1300_InstantSensorData}
\alias{DataMap_Lotek_1300_InstantSensorData}
\title{DataMap - Instant sensor data from Lotek 1300 tags}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_Lotek_1300_InstantSensorData
}
\description{
DataMap - Instant sensor data from Lotek 1300 tags
}
\keyword{datasets}
