% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_Lotek_1300_TagMetaData}
\alias{DataMap_Lotek_1300_TagMetaData}
\title{DataMap - Tag metadata from Lotek 1300 tags}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_Lotek_1300_TagMetaData
}
\description{
DataMap - Tag metadata from Lotek 1300 tags
}
\keyword{datasets}
