% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_Lotek_1400.1800_InstantSensorData}
\alias{DataMap_Lotek_1400.1800_InstantSensorData}
\title{Datamap - Lotek 1400/1800 instant sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_Lotek_1400.1800_InstantSensorData
}
\description{
Datamap - Lotek 1400/1800 instant sensor data
}
\keyword{datasets}
