% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_Lotek_1400.1800_TagMetaData}
\alias{DataMap_Lotek_1400.1800_TagMetaData}
\title{DataMap - Tag metadata from Lotek 1400/1800 tags}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_Lotek_1400.1800_TagMetaData
}
\description{
DataMap - Tag metadata from Lotek 1400/1800 tags
}
\keyword{datasets}
