% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek_InstantSensorData-class}
\alias{DataMap_Lotek_InstantSensorData-class}
\alias{DataMap_Lotek_InstantSensorData}
\title{DataMap - Base map for Lotek instant sensor data}
\description{
DataMap - Base map for Lotek instant sensor data
}

