% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_Lotek_TagMetaData-class}
\alias{DataMap_Lotek_TagMetaData-class}
\alias{DataMap_Lotek_TagMetaData}
\title{DataMap - Lotek tag metadata - base class}
\description{
DataMap - Lotek tag metadata - base class
}

