% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_MicrowaveTelemetry_TagMetaData-class}
\alias{DataMap_MicrowaveTelemetry_TagMetaData-class}
\alias{DataMap_MicrowaveTelemetry_TagMetaData}
\title{DataMap - Microwave Telemetry tag metadata - base class}
\description{
DataMap - Microwave Telemetry tag metadata - base class
}

