% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_MicrowaveTelemetry_XTag_InstantSensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_InstantSensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_InstantSensorData}
\title{Datamap - Microwave Telemetry X-tag instant sensor data}
\description{
Datamap - Microwave Telemetry X-tag instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

