% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_MicrowaveTelemetry_XTag_SensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_SensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_SensorData}
\title{DataMap - Microwave Telemetry X-Tag metadata}
\description{
DataMap - Microwave Telemetry X-Tag metadata
}

