% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_MicrowaveTelemetry_XTag_SummarySensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_SummarySensorData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_SummarySensorData}
\title{Datamap - Microwave Telemetry X-tag summary sensor data}
\description{
Datamap - Microwave Telemetry X-tag summary sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

