% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_MicrowaveTelemetry_XTag_SummarySensorData}
\alias{DataMap_MicrowaveTelemetry_XTag_SummarySensorData}
\title{Datamap - Microwave Telemetry X-tag summary sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_MicrowaveTelemetry_XTag_SummarySensorData
}
\description{
Datamap - Microwave Telemetry X-tag summary sensor data
}
\keyword{datasets}
