% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_MicrowaveTelemetry_XTag_TagMetaData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_TagMetaData-class}
\alias{DataMap_MicrowaveTelemetry_XTag_TagMetaData}
\title{DataMap - Microwave Telemetry X-Tag metadata}
\description{
DataMap - Microwave Telemetry X-Tag metadata
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

