% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_MicrowaveTelemetry_XTag_Transmitted_InstantSensorData}
\alias{DataMap_MicrowaveTelemetry_XTag_Transmitted_InstantSensorData}
\title{Datamap - Microwave Telemetry X-tag instant sensor data (transmitted via Satellite)}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_MicrowaveTelemetry_XTag_Transmitted_InstantSensorData
}
\description{
Datamap - Microwave Telemetry X-tag instant sensor data (transmitted via Satellite)
}
\keyword{datasets}
