% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_MicrowaveTelemetry_XTag_Transmitted_TagMetaData}
\alias{DataMap_MicrowaveTelemetry_XTag_Transmitted_TagMetaData}
\title{DataMap - Microwave Telemetry X-Tag metadata (transmitted via satellite)}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_MicrowaveTelemetry_XTag_Transmitted_TagMetaData
}
\description{
DataMap - Microwave Telemetry X-Tag metadata (transmitted via satellite)
}
\keyword{datasets}
