% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_StarOddi_DST_InstantSensorData-class}
\alias{DataMap_StarOddi_DST_InstantSensorData-class}
\alias{DataMap_StarOddi_DST_InstantSensorData}
\title{DataMap - StarOddi DST instant sensor data}
\description{
DataMap - StarOddi DST instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

