% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_StarOddi_DST_InstantSensorData}
\alias{DataMap_StarOddi_DST_InstantSensorData}
\title{DataMap - StarOddi DST instant sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_StarOddi_DST_InstantSensorData
}
\description{
DataMap - StarOddi DST instant sensor data
}
\keyword{datasets}
