% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_StarOddi_DST_TagMetaData-class}
\alias{DataMap_StarOddi_DST_TagMetaData-class}
\alias{DataMap_StarOddi_DST_TagMetaData}
\title{DataMap - StarOddi DST tag metadata}
\description{
DataMap - StarOddi DST tag metadata
}

