% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_StarOddi_DST_TagMetaData}
\alias{DataMap_StarOddi_DST_TagMetaData}
\title{DataMap - StarOddi tag metadata - base class}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_StarOddi_DST_TagMetaData
}
\description{
DataMap - StarOddi tag metadata - base class
}
\keyword{datasets}
