% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_StarOddi_DSTmagnetic_TagMetaData-class}
\alias{DataMap_StarOddi_DSTmagnetic_TagMetaData-class}
\alias{DataMap_StarOddi_DSTmagnetic_TagMetaData}
\title{DataMap - StarOddi DST magnetic tag metadata}
\description{
DataMap - StarOddi DST magnetic tag metadata
}

