% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_StarOddi_DSTmagnetic_TagMetaData}
\alias{DataMap_StarOddi_DSTmagnetic_TagMetaData}
\title{DataMap - StarOddi DST magnetic tag metadata}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_StarOddi_DSTmagnetic_TagMetaData
}
\description{
DataMap - StarOddi DST magnetic tag metadata
}
\keyword{datasets}
