% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_StarOddi_TagMetaData-class}
\alias{DataMap_StarOddi_TagMetaData-class}
\alias{DataMap_StarOddi_TagMetaData}
\title{DataMap - StarOddi tag metadata - base class}
\description{
DataMap - StarOddi tag metadata - base class
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

