% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_SummarySensorData_Base-class}
\alias{DataMap_SummarySensorData_Base-class}
\alias{DataMap_SummarySensorData_Base}
\title{DataMap - Summary sensor data - base class}
\description{
DataMap - Summary sensor data - base class
}

