% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\docType{class}
\name{DataMap_TagMetaData-class}
\alias{DataMap_TagMetaData-class}
\alias{DataMap_TagMetaData}
\title{Tag metadata map}
\description{
Base class for tag metadata DataMaps.
}
\section{Fields}{

\describe{
\item{\code{make}}{character. Tag manufacturer/brand}

\item{\code{model}}{character. Tag model}
}}

\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

