% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_TagMetaData_Base-class}
\alias{DataMap_TagMetaData_Base-class}
\alias{DataMap_TagMetaData_Base}
\title{Tag metadata map}
\description{
Base class for DataMaps for tag metadata
}
\section{Fields}{

\describe{
\item{\code{make}}{character. Tag manufacturer/brand}

\item{\code{model}}{character. Tag model}
}}

\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}

\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

