% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_WildlifeComputer_BenthicSPAT_SummarySensorData-class}
\alias{DataMap_WildlifeComputer_BenthicSPAT_SummarySensorData-class}
\alias{DataMap_WildlifeComputer_BenthicSPAT_SummarySensorData}
\title{DataMap - Wildlife Computers Benthic sPAT summary sensor data}
\description{
DataMap - Wildlife Computers Benthic sPAT summary sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

