% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_WildlifeComputer_MiniPAT_SummarySensorData-class}
\alias{DataMap_WildlifeComputer_MiniPAT_SummarySensorData-class}
\alias{DataMap_WildlifeComputer_MiniPAT_SummarySensorData}
\title{DataMap - Wildlife Computers MiniPAT summary sensor}
\description{
DataMap - Wildlife Computers MiniPAT summary sensor
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

