% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData-class}
\alias{DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData-class}
\alias{DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData}
\title{DataMap - Wildlife Computers Benthic sPAT instant sensor data}
\description{
DataMap - Wildlife Computers Benthic sPAT instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

