% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData}
\alias{DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData}
\title{DataMap - Wildlife Computers Benthic sPAT instant sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_BenthicSPAT_InstantSensorData
}
\description{
DataMap - Wildlife Computers Benthic sPAT instant sensor data
}
\keyword{datasets}
