% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_WildlifeComputers_BenthicSPAT_SummarySensorData}
\alias{DataMap_WildlifeComputers_BenthicSPAT_SummarySensorData}
\title{DataMap - Wildlife Computers Benthic sPAT summary sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_BenthicSPAT_SummarySensorData
}
\description{
DataMap - Wildlife Computers Benthic sPAT summary sensor data
}
\keyword{datasets}
