% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_WildlifeComputers_BenthicSPAT_TagMetaData-class}
\alias{DataMap_WildlifeComputers_BenthicSPAT_TagMetaData-class}
\alias{DataMap_WildlifeComputers_BenthicSPAT_TagMetaData}
\title{DataMap - Wildlife Computers Benthic sPAT tag metadata}
\description{
DataMap - Wildlife Computers Benthic sPAT tag metadata
}
\section{Methods}{

\describe{
\item{\code{get_tag_id(d)}}{Extract tag id from directory `d`}
}}

