% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_WildlifeComputers_BenthicSPAT_TagMetaData}
\alias{DataMap_WildlifeComputers_BenthicSPAT_TagMetaData}
\title{DataMap - Wildlife Computers Benthic sPAT tag metadata}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_WildlifeComputers_BenthicSPAT_TagMetaData
}
\description{
DataMap - Wildlife Computers Benthic sPAT tag metadata
}
\keyword{datasets}
