% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_WildlifeComputers_MiniPAT_HistogramMetaData}
\alias{DataMap_WildlifeComputers_MiniPAT_HistogramMetaData}
\title{DataMap - Wildlife Computers MiniPAT histogram data metadata}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_MiniPAT_HistogramMetaData
}
\description{
DataMap - Wildlife Computers MiniPAT histogram data metadata
}
\keyword{datasets}
