% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_WildlifeComputers_MiniPAT_InstantSensorData}
\alias{DataMap_WildlifeComputers_MiniPAT_InstantSensorData}
\title{DataMap - Wildlife Computers MiniPAT instant sensor data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_MiniPAT_InstantSensorData
}
\description{
DataMap - Wildlife Computers MiniPAT instant sensor data
}
\keyword{datasets}
