% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.R
\docType{data}
\name{DataMap_WildlifeComputers_MiniPAT_PDTData}
\alias{DataMap_WildlifeComputers_MiniPAT_PDTData}
\title{DataMap - Wildlife Computers MiniPAT PDT (Profile of Depth and Temperature) data}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_MiniPAT_PDTData
}
\description{
DataMap - Wildlife Computers MiniPAT PDT (Profile of Depth and Temperature) data
}
\keyword{datasets}
