% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_WildlifeComputers_MiniPAT_SummarySensorData}
\alias{DataMap_WildlifeComputers_MiniPAT_SummarySensorData}
\title{DataMap - Wildlife Computers MiniPAT summary sensor}
\format{
An object of class \code{DataMap} of length 1.
}
\usage{
DataMap_WildlifeComputers_MiniPAT_SummarySensorData
}
\description{
DataMap - Wildlife Computers MiniPAT summary sensor
}
\keyword{datasets}
