% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{data}
\name{DataMap_WildlifeComputers_MiniPAT_TagMetaData}
\alias{DataMap_WildlifeComputers_MiniPAT_TagMetaData}
\title{DataMap - Wildlife Computers tag metadata - base class}
\format{
An object of class \code{DataMap_TagMetaData} of length 1.
}
\usage{
DataMap_WildlifeComputers_MiniPAT_TagMetaData
}
\description{
DataMap - Wildlife Computers tag metadata - base class
}
\keyword{datasets}
