% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{DataMap_WildlifeComputers_TagMetaData-class}
\alias{DataMap_WildlifeComputers_TagMetaData-class}
\alias{DataMap_WildlifeComputers_TagMetaData}
\title{DataMap - Wildlife Computers tag metadata - base class}
\description{
DataMap - Wildlife Computers tag metadata - base class
}

