% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Decoder-class}
\alias{Decoder-class}
\alias{Decoder}
\title{Decoder base class}
\arguments{
\item{dat_1}{The dataframe being altered}

\item{dm_1}{DataMap of the dataframe being altered}

\item{dat_2}{Dataframe from which data may be taken}

\item{dm_2}{DataMap of dataframe from which data may be taken}

\item{dm}{The DataMap to use}

\item{con}{Connection to the target DB}
}
\value{
dat_1 with any fields which were missing and were present in dat_2
Execute all necessary steps to read and transform raw data for one DataMap

The extracted and transformed data
Execute all etl steps for all DataMaps

Boolean value indicating if the tag ID is already present in the metadata table
Execute all necessary steps to read and transform raw data for one datamap

The extracted and transformed data
}
\description{
Logical representation of a single tag
}
\section{Fields}{

\describe{
\item{\code{d}}{character. The directory in which the tag data can be found}

\item{\code{data_maps}}{list. The list of data maps used to map the data for this tag to the output tables in the DB

The following fields should be set in initialization of child classes}

\item{\code{tag_make}}{character. The manufacturer of the tag}

\item{\code{tag_model}}{character. The model of the tag}
}}


\examples{
Add fields that are missing from a data frame

Take in two data frames and their respective DataMaps
Any fields which are missing from the first dataframe and present in the second dataframe are added to the first

}
