% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder_config.R
\docType{class}
\name{Decoder_Lotek-class}
\alias{Decoder_Lotek-class}
\alias{Decoder_Lotek}
\title{Base implementation for all Lotek tag Decoders}
\description{
Base implementation for all Lotek tag Decoders
}

