% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder_config.R
\docType{class}
\name{Decoder_MicrowaveTelemetry-class}
\alias{Decoder_MicrowaveTelemetry-class}
\alias{Decoder_MicrowaveTelemetry}
\title{Base implementation for all Microwave Telemetry tag Decoders}
\description{
Base implementation for all Microwave Telemetry tag Decoders
}

