% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datamap_config.r
\docType{class}
\name{Decoder_StarOddi_DSTmagnetic_InstantSensorData-class}
\alias{Decoder_StarOddi_DSTmagnetic_InstantSensorData-class}
\alias{DataMap_StarOddi_DSTmagnetic_InstantSensorData}
\title{DataMap - StarOddi DST magnetic instant sensor data}
\description{
DataMap - StarOddi DST magnetic instant sensor data
}
\section{Methods}{

\describe{
\item{\code{extract(d)}}{Extract data from directory `d`}
}}

