% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{DESERTSTAR_SEATAG_MOD_INSTANT_DATA_FIELDS}
\alias{DESERTSTAR_SEATAG_MOD_INSTANT_DATA_FIELDS}
\title{DesertStar SeaTag MOD data fields}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
DESERTSTAR_SEATAG_MOD_INSTANT_DATA_FIELDS
}
\description{
Unlke most input data FieldMap objects, the Field objects in this FieldMap have their data_types specified. This is because, unlike the other tag-types, Decoding DesertStar data is a two-step process, so this input FieldMap will initially be used as both the input AND output FieldMap object, and will therefore be checked to ensure that the output values match the expected types.
}
\keyword{datasets}
