% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Field-class}
\alias{Field-class}
\alias{Field}
\title{Field class. Represents a single data field and its key metadata for future}
\description{
Field class. Represents a single data field and its key metadata for future
}
\section{Fields}{

\describe{
\item{\code{name}}{character. The name of the field}

\item{\code{units}}{character. The SI units of the field}

\item{\code{data_type}}{character. The data type of this field in a DB}
}}


