% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{FieldInputForm-class}
\alias{FieldInputForm-class}
\alias{FieldInputForm}
\title{Construct forms for user-input data}
\description{
The \code{FieldInputForm} class provides an interface by which we can quickly create input forms to prompt users for missing data. \code{FieldInputForm} provides methods with which to construct pop-up windows from a list of \code{InputField} object, and then to retrieve the data submitted by the user for each \code{InputField}.
}
\section{Fields}{

\describe{
\item{\code{id_fields}}{character.}

\item{\code{window_width}}{numeric.}
}}


