% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{FieldValidator-class}
\alias{FieldValidator-class}
\alias{FieldValidator}
\title{Validate user-input fields}
\description{
Provides a simple interface for defining validation rules. Each FieldValidator object contains a single function \code{validate} which accepts a single value \code{val} and returns a boolean value indicating if that value is valid or not. If not, then the error message will be used to inform the user of the discrepency.
}
\section{Fields}{

\describe{
\item{\code{validate}}{function. Accepts a single parameter \code{val}, returns \code{TRUE} if \code{val} is valid for the given field.}

\item{\code{error_msg}}{character. Message used to inform the user of the invalid value and how it should be changed}
}}


