% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_util.R
\name{files_by_pattern}
\alias{files_by_pattern}
\title{Search for one (or more) files in a directory by pattern}
\usage{
files_by_pattern(d, pattern, n = 1, ...)
}
\arguments{
\item{d}{The directory to search in}

\item{pattern}{Regex pattern to match against filenames}

\item{n}{The number of files expected to be found. Default is 1}
}
\value{
All matching filenames
}
\description{
Searches directory \code{d} for files which match the given pattern, and checks to ensure that the number of matching files matches the expected number of files. If not, an error is thrown, otherwise the matching filename(s) is(are) returned
}
