% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_cond_stack_messages}
\alias{get_cond_stack_messages}
\title{Collect and return the messages from a condition stack}
\usage{
get_cond_stack_messages(cond, l = list())
}
\arguments{
\item{cond}{The condition from which to begin traversing the stack (from child to parent)}
}
\value{
A list of messages from the condition stack, including offending calls
}
\description{
Collect and return the messages from a condition stack
}
