% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{get_field_descriptions}
\alias{get_field_descriptions}
\title{Compile a hierarchical list of field descriptions from a list of output fieldmaps}
\usage{
get_field_descriptions(output_fieldmaps)
}
\arguments{
\item{output_fieldmaps}{A list of output fieldmaps}
}
\value{
A hierarchical list of field descriptions
}
\description{
Compile a hierarchical list of field descriptions from a list of output fieldmaps
}
