% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Identifier-class}
\alias{Identifier-class}
\alias{Identifier}
\title{Identifier class. Used to identify if a given directory belongs to a particular make/model of tag, based on available metadata}
\description{
Base class. Not intended to be implemented directly.
}

