% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{class}
\name{Identifier_DesertStar_SeaTagMOD-class}
\alias{Identifier_DesertStar_SeaTagMOD-class}
\alias{Identifier_DesertStar_SeaTagMOD}
\title{Identifier - Desert Star SeaTag MOD tags}
\description{
Identifier - Desert Star SeaTag MOD tags
}
\section{Methods}{

\describe{
\item{\code{identify(d)}}{Returns TRUE/FALSE indicating if the metadata in the passed directory
matches the patterns expected of the tag type with which this Identifier
child class is associated}
}}

