% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_DesertStar_SeaTagMOD}
\alias{Identifier_DesertStar_SeaTagMOD}
\title{Identifier - Desert Star SeaTag MOD tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_DesertStar_SeaTagMOD
}
\description{
Identifier - Desert Star SeaTag MOD tags
}
\keyword{datasets}
