% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_DesertStar_SeaTagMOD_DIRECTORY}
\alias{Identifier_DesertStar_SeaTagMOD_DIRECTORY}
\title{Identifier - Desert Star SeaTag MOD tags - Directory of DesertStar raw datafiles (containing data from many tags)}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_DesertStar_SeaTagMOD_DIRECTORY
}
\description{
Identifier - Desert Star SeaTag MOD tags - Directory of DesertStar raw datafiles (containing data from many tags)
}
\keyword{datasets}
