% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_DesertStar_SeaTagMOD_SINGLETAG}
\alias{Identifier_DesertStar_SeaTagMOD_SINGLETAG}
\title{Identifier - Desert Star SeaTag MOD tags - Directory containing data from a single tag}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_DesertStar_SeaTagMOD_SINGLETAG
}
\description{
Identifier - Desert Star SeaTag MOD tags - Directory containing data from a single tag
}
\keyword{datasets}
