% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_Lotek_1000.1100.1250}
\alias{Identifier_Lotek_1000.1100.1250}
\title{Identifier - Lotek 1000/1100/1250 tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_Lotek_1000.1100.1250
}
\description{
Identifier - Lotek 1000/1100/1250 tags
}
\keyword{datasets}
