% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{class}
\name{Identifier_Lotek_1300-class}
\alias{Identifier_Lotek_1300-class}
\alias{Identifier_Lotek_1300}
\title{Identifier - Lotek 1300 tags}
\description{
Identifier - Lotek 1300 tags
}
\section{Methods}{

\describe{
\item{\code{identify(d)}}{Returns TRUE/FALSE indicating if the metadata in the passed directory
matches the patterns expected of the tag type with which this Identifier
child class is associated}
}}

