% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_Lotek_1300}
\alias{Identifier_Lotek_1300}
\title{Identifier - Lotek 1300 tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_Lotek_1300
}
\description{
Identifier - Lotek 1300 tags
}
\keyword{datasets}
