% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{class}
\name{Identifier_Lotek_1400.1800-class}
\alias{Identifier_Lotek_1400.1800-class}
\alias{Identifier_Lotek_1400.1800}
\title{Identifier - Lotek 1400/1800 tags}
\description{
Identifier - Lotek 1400/1800 tags
}
\section{Methods}{

\describe{
\item{\code{identify(d)}}{Returns TRUE/FALSE indicating if the metadata in the passed directory
matches the patterns expected of the tag type with which this Identifier
child class is associated}
}}

