% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_Lotek_1400.1800}
\alias{Identifier_Lotek_1400.1800}
\title{Identifier - Lotek 1400/1800 tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_Lotek_1400.1800
}
\description{
Identifier - Lotek 1400/1800 tags
}
\keyword{datasets}
