% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_MicrowaveTelemetry_XTag}
\alias{Identifier_MicrowaveTelemetry_XTag}
\title{Identifier - Microwave Telemetry X-tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_MicrowaveTelemetry_XTag
}
\description{
Identifier - Microwave Telemetry X-tags
}
\keyword{datasets}
