% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_MicrowaveTelemetry_XTag_Recovered}
\alias{Identifier_MicrowaveTelemetry_XTag_Recovered}
\title{Identifier - Microwave Telemetry X-tags (physically recovered)}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_MicrowaveTelemetry_XTag_Recovered
}
\description{
Identifier - Microwave Telemetry X-tags (physically recovered)
}
\keyword{datasets}
