% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_MicrowaveTelemetry_XTag_Transmitted}
\alias{Identifier_MicrowaveTelemetry_XTag_Transmitted}
\title{Identifier - Microwave Telemetry X-tags (transmitted via satellite)}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_MicrowaveTelemetry_XTag_Transmitted
}
\description{
Identifier - Microwave Telemetry X-tags (transmitted via satellite)
}
\keyword{datasets}
