% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{class}
\name{Identifier_StarOddi-class}
\alias{Identifier_StarOddi-class}
\alias{Identifier_StarOddi}
\title{Identifier - StarOddi tags - base class}
\description{
Identifier - StarOddi tags - base class
}

