% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_StarOddi_DST}
\alias{Identifier_StarOddi_DST}
\title{Identifier - StarOddi DST tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_StarOddi_DST
}
\description{
Identifier - StarOddi DST tags
}
\keyword{datasets}
