% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{class}
\name{Identifier_WildlifeComputers-class}
\alias{Identifier_WildlifeComputers-class}
\alias{Identifier_WildlifeComputers}
\title{Identifier - Wildlife Computer tags}
\description{
Identifier - Wildlife Computer tags
}

