% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_WildlifeComputers_BenthicSPAT}
\alias{Identifier_WildlifeComputers_BenthicSPAT}
\title{Identifier - Wildlife Computer Benthic sPAT tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_WildlifeComputers_BenthicSPAT
}
\description{
Identifier - Wildlife Computer Benthic sPAT tags
}
\keyword{datasets}
