% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identifier_config.R
\docType{data}
\name{Identifier_WildlifeComputers_MiniPAT}
\alias{Identifier_WildlifeComputers_MiniPAT}
\title{Identifier - Wildlife Computer MiniPAT tags}
\format{
An object of class \code{Identifier} of length 1.
}
\usage{
Identifier_WildlifeComputers_MiniPAT
}
\description{
Identifier - Wildlife Computer MiniPAT tags
}
\keyword{datasets}
