% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{IndependentField-class}
\alias{IndependentField-class}
\alias{IndependentField}
\title{IndependentField}
\description{
A Field object which generates its own data. Useful for autogenerated timestamps and such.
}
\section{Fields}{

\describe{
\item{\code{value_fn}}{function. Accepts \code{dat} and any number of named arguments. Returns the value for this field}
}}


