% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{InputField-class}
\alias{InputField-class}
\alias{InputField}
\title{InputField - base class}
\description{
Base of all user input fields. Not intended to be directly implemented, use a child class instead
}
\section{Methods}{

\describe{
\item{\code{build_label(window, ...)}}{Build an appropriate label for the input widget}

\item{\code{build_widget(window, ...)}}{Return a fully constructed and configured widget for use in the input form}

\item{\code{get_validation_errors(val, ...)}}{Get all of the validation errors triggered by `val`}

\item{\code{get_value()}}{Retrieve the user-inputted value for this field}

\item{\code{register_tclvar(var_label, init_value = "")}}{Register a new tclvar object}

\item{\code{register_validator(validator)}}{Add a validator to this Field object}

\item{\code{validate_value(val, ...)}}{Returns a boolean indicating if `val` is valid or not.}
}}

