% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{InputField_Date-class}
\alias{InputField_Date-class}
\alias{InputField_Date}
\title{InputField object for selecting dates}
\description{
InputField object for selecting dates
}
\section{Methods}{

\describe{
\item{\code{build_widget(window, ...)}}{Return a fully constructed and configured widget for use in the input form}

\item{\code{get_value()}}{Retrieve the user-inputted value for this field}
}}

