% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{InputField_FilteredSelect-class}
\alias{InputField_FilteredSelect-class}
\alias{InputField_FilteredSelect}
\title{Expanded version of \code{InputField_Select} which offers filtering functionality}
\description{
Operates identically to \code{InputField_Select}, except that when the list of choices is retrieved from \code{table}, a filter is applied to reduce the number of options available to the user. \code{table} is first filtered to only include records for which \code{filter_db_field} values are equal to the value contained in the \code{filter_input_field} in the incoming data. \code{choices} and \code{pks} are then populated from this filtered set of records in the normal way.
}
\section{Fields}{

\describe{
\item{\code{filter_db_field}}{character. Name of field in \code{table} which will be used to filter options}

\item{\code{filter_input_field}}{character. Name of field in incoming data which will contain the value used to filter \code{table} (i.e. incoming_data[\link{filter_input_field}] == table[\link{filter_db_field}])}
}}


