% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{InputField_Select-class}
\alias{InputField_Select-class}
\alias{InputField_Select}
\title{An input field for which the user selects from a list of available options}
\description{
Useful when we want to restrict the possible inputs a user can provide, for example if they are choosing the species. Rather than asking them to retype the species every time, each time potentially yielding a mispelling or other mistake we provide them with a list of options from which they can choose.
}
\section{Fields}{

\describe{
\item{\code{table}}{character. Name of DB table from which the choices will be collected}

\item{\code{choice_field}}{character. Field in \code{table} which contains the choices to offer}

\item{\code{pk_field}}{character. Field in \code{table} which contains the primary keys of the choices}
}}

\section{Methods}{

\describe{
\item{\code{build_widget(window, ...)}}{Return a fully constructed and configured widget for use in the input form}

\item{\code{get_value()}}{Retrieve the user-inputted value for this field}
}}

