% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{InputField_SelectAdd-class}
\alias{InputField_SelectAdd-class}
\alias{InputField_SelectAdd}
\title{A selection input field capable of adding new entries}
\description{
An expanded version of the InputField_Select class. Functions identically, except that when a value not currently present in the table is supplied by the user, automatically tries to insert the new value into \code{choice_field} of \code{table}. Assumes that \code{table} is capable of auto-generating primary key values for new entries. If this is not setup, behavior is undefined.
}
\section{Methods}{

\describe{
\item{\code{get_value(...)}}{Retrieve the user-inputted value for this field}
}}

