% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{names_rec}
\alias{names_rec}
\title{Recursively traverse nested list, printing out list names}
\usage{
names_rec(l, current_depth = 1, max_depth = 1)
}
\arguments{
\item{l}{list to traverse}

\item{max_depth}{maximum depth of recursion}
}
\description{
Recursively traverse nested list, printing out list names
}
