% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\docType{class}
\name{Pipe-class}
\alias{Pipe-class}
\alias{Pipe}
\title{Pipe class.}
\description{
Facilitates the automated identification and decoding of
tag data directories. Recursively traverses a directory tree, attempting to
process any data directories within.
}
\details{
Implements some basic reporting to aid in diagnosing systemic decoding issues
}
\section{Fields}{

\describe{
\item{\code{d}}{Data directory.}

\item{\code{decoders}}{A list of Decoder objects which can be use to extract and load the data produced by different tags}

\item{\code{overwrite}}{Boolean flag. If set to TRUE, the Pipe will check if a tag is already present in the tag metadata table, and will not process the tag if it is.}
}}

\section{Methods}{

\describe{
\item{\code{add_missing_fields(dat1, dat1_ip_fm, dat1_op_fm, dat2, dat2_ip_fm, dat2_op_fm)}}{Add fields from one data frame to a second dataframe, based on their
respective DataMaps. dat_1/dm_1 refer to the data which may be missing
fields. dat_2/dm_2 refer to the data from which we MAY take data to
put into dat_1. Any fields defined by dm_1$output_data_field_map which
are missing from dat_1 but are present in dat_2 will be added to dat_1.}

\item{\code{build_report()}}{Builds a report (data.frame) detailing the current state of the data directory}

\item{\code{pre_process_node(node)}}{Perform any necessary pre-processing on a directory}

\item{\code{pre_process_nodes()}}{Perform any necessary pre-processing on the nodes of the directory tree}

\item{\code{process_node(node, con, overwrite = T)}}{Attempt to process directory `d`. Pipe will first attempt to identify the make/model of the source tag, and if one is found will apply the corresponding decoder to the directory to extract, transform, and load the data.}

\item{\code{process_to_csv(out_d, ...)}}{Extract tag data and write it to csv files in `out_d`}

\item{\code{process_to_dataframes(con, ...)}}{Extract tag data and return it as a list of dataframes}

\item{\code{process_to_db(con, overwrite = T, silent = F)}}{Recursively process a directory tree, attempting to process each directory.}

\item{\code{refresh_op_fm(...)}}{Refresh each of the FieldMap objects}

\item{\code{register_decoder(dc)}}{Add one or more Decoder objects to the list of decoders used by this Pipe}
}}

\examples{
data_directory = here::here() # Root data directory

# Initialize the object
p__ = Pipe(d = data_directory)

# Process the directory
p__$process(con = con, overwrite = F)

# Build the processing report
report = p__$build_report()
print(report)
}
