% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{posix_timestamp_to_character}
\alias{posix_timestamp_to_character}
\title{Convert POSIXct timestamp to character string}
\usage{
posix_timestamp_to_character(v)
}
\arguments{
\item{v}{A vector of POSIXct timestamps}
}
\description{
Provides a way to convert a POSIXct timestamp to a character string in a consistant manner
}
