% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{prepend_error_message}
\alias{prepend_error_message}
\title{Prepend error message}
\usage{
prepend_error_message(cond, prefix)
}
\arguments{
\item{cond}{The condition to which to prepend the message}

\item{prefix}{The message to prepend}
}
\description{
Utility function to prepend messages to error conditions. Useful for adding additional details about when, where, and how an error occurred
}
