% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{TagProcessor-class}
\alias{TagProcessor-class}
\alias{TagProcessor}
\title{TagProcessor class.}
\value{
Instance of the TagProcessor class
}
\description{
Facilitates the automated identification and decoding of
tag data directories. Recursively traverses a directory tree, attempting to
process any data directories within.
}
\details{
Implements some basic reporting to aid in diagnosing systemic decoding issues
}
\section{Fields}{

\describe{
\item{\code{d}}{Data directory.}

\item{\code{overwrite}}{Boolean flag. If set to TRUE, the TagProcessor will check if a tag is already present in the tag metadata table, and will not process the tag if it is.}
}}

\section{Methods}{

\describe{
\item{\code{build_report()}}{Builds a report (data.frame) detailing the current state of the data directory}

\item{\code{process(con, overwrite = T)}}{Recursively process a directory tree, attempting to process each directory.}

\item{\code{process_node(node, con, overwrite = T)}}{Attempt to process directory `d`. TagProcessor will first attempt to identify the make/model of the source tag, and if one is found will apply the corresponding decoder to the directory to extract, transform, and load the data.}
}}

\examples{
data_directory = here::here() # Root data directory

# Initialize the object
tp__ = TagProcessor(d = data_directory)

# Process the directory
tp__$process(con = con, overwrite = F)

# Build the processing report
report = tp__$build_report()
print(report)
}
