% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field_config.R
\docType{data}
\name{USER_INPUT_FIELDS}
\alias{USER_INPUT_FIELDS}
\title{Fields provided by the user}
\format{
An object of class \code{FieldMap} of length 1.
}
\usage{
USER_INPUT_FIELDS
}
\description{
Fields provided by the user
}
\keyword{datasets}
