% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_util.R
\name{wc_minipat_calc_end_times}
\alias{wc_minipat_calc_end_times}
\title{Calculate end times for Wildlife Computers MiniPAT histogram and profile data}
\usage{
wc_minipat_calc_end_times(start_times)
}
\arguments{
\item{start_times}{A vector of POSIXct objects representing the start times of the records, as they are recorded in the input dataframe}
}
\value{
A vector of POSIXct objects representing the endtimes of the records, calculated based on the start times
}
\description{
WC MiniPAT tags record histogram and profile data as a series of records, each of which has a start time. The end time of each record is not explicitly recorded, but can be inferred from the start time of the following record. This function calculates the end time for each record in a series of start times.
}
