% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_util.R
\name{wc_minipat_find_interval}
\alias{wc_minipat_find_interval}
\title{Determine interval from vector of dates}
\usage{
wc_minipat_find_interval(date_)
}
\arguments{
\item{date_}{Vector of POSIXct dates}
}
\value{
The interval at which the dates were programmed to be spaced, expressed as a difftime obj
}
\description{
Wildlife Computer MiniPAT tags collect certain data continuously, but rather than report the continuous data stream they report a series of packages, each of which contains the aggregate stats for a certain interval of time. This interval is not explicitly defined in the output data, but can be inferred from the timestamps of the data. This function performs that inference
}
