% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_map.R
\docType{class}
\name{DataMap-class}
\alias{DataMap-class}
\alias{DataMap}
\title{DataMap}
\value{
An instance of the DataMap class
}
\description{
Provides base structure for all make/model specific tag data maps
Analagous to a table in the final DB. Desribes programatically how the data
to be inserted into that table should be collected from the disk, and
transformed.
Not intended to be implemented directly
}
\section{Fields}{

\describe{
\item{\code{input_data_field_map}}{FieldMap. Map of the data fields and their original format}

\item{\code{extract_fn}}{function(d) Function which extracts the appropriate data from the files in directory \code{d}}
}}

\section{Methods}{

\describe{
\item{\code{augment(dat, output_data_field_map)}}{Augment the transformed dataset with any missing data}

\item{\code{extract(d)}}{Extract data from directory `d`}

\item{\code{transform(dat, output_data_field_map)}}{Transform extracted data, as dictated by the input/output `FieldMap` objects}
}}

