% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decoder.R
\docType{class}
\name{Decoder-class}
\alias{Decoder-class}
\alias{Decoder}
\title{Decoder base class}
\description{
Logical representation of a single tag, which contains all necessary logic to parse and extract all required data from an appropriately formatted directory. Contains one or more \link{DataMap} objects, each of which the Decoder will apply in turn to the directory in order to extract the necessary data. Contains an \link{Identifier} object, which can be used to automatically determine if a given directory contains the data generated by a tag of the make/model to which this Decoder refers.
}
\section{Fields}{

\describe{
\item{\code{data_maps}}{list. The list of data maps used to map the data for this tag to the output tables in the DB}
}}

\section{Methods}{

\describe{
\item{\code{decode_datamap(d, dm, op_fm)}}{Execute all necessary steps to read and transform raw data for one DataMap}
}}

