% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/field.R
\docType{class}
\name{FieldMap-class}
\alias{FieldMap-class}
\alias{FieldMap}
\title{FieldMap class.}
\description{
Collection of Field objects, with some added functionality to facilitate the convenient use of said Field objects both within the package and when interfacing with a DB
}
\section{Fields}{

\describe{
\item{\code{fields}}{Field. The Field objects contained within this map/}
}}

\section{Methods}{

\describe{
\item{\code{common_fields(fm)}}{Generates a list of the Fields shared by this FieldMap and the given FieldMap}

\item{\code{get_field_names()}}{Get the names of the ID fields in this FieldMap}

\item{\code{get_id_field_names()}}{Get the names of the ID fields in this FieldMap}

\item{\code{get_id_fields()}}{Return the subset of Fields in this FieldMap that are marked as ID fields}

\item{\code{get_independent_fields()}}{Get all independent Field objects as a named list}

\item{\code{uncommon_fields(fm)}}{Generate a list of the Fields NOT shared by this FieldMap and the given FieldMap}
}}

