% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Identifier-class}
\alias{Identifier-class}
\alias{Identifier}
\title{Identifier class. Used to identify if a given directory belongs to a particular make/model of tag, based on available metadata}
\description{
Base class. Not intended to be implemented directly.
Returns TRUE/FALSE indicating if the metadata in the passed directory matches the patterns expected of the tag type with which this Identifier child class is associated
Check number of files in directory which match pattern
}
\details{
Helper function to quickly determine if an expected number of files
match the given pattern
}

