% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{TagIdentifier-class}
\alias{TagIdentifier-class}
\alias{TagIdentifier}
\title{Maps raw tag data to the appropriate Decoder based on structure of the raw data}
\description{
Makes decisions based on some pretty nitty-gritty details, like naming
conventions, file-types present, number of files present, etc. As such, it's
crucial that the data has not been altered by the user in any way since being
extracted from the tag and/or run through any post-processing software.
}
\section{Fields}{

\describe{
\item{\code{master_list__}}{list. The list of all Decoders from which to choose.
Internal reference field, not intended to be set by user. Any value passed
to this field will be overwritten on construction}
}}


