% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{TagProcessor-class}
\alias{TagProcessor-class}
\alias{TagProcessor}
\title{TagProcessor class.}
\value{
Instance of the TagProcessor class
}
\description{
Facilitates the automated identification and decoding of
tag data directories. Recursively traverses a directory tree, attempting to
process any data directories within.
}
\details{
Implements some basic reporting to aid in diagnosing systemic decoding issues
}
\section{Fields}{

\describe{
\item{\code{d}}{Data directory.}

\item{\code{overwrite}}{Boolean flag. If set to TRUE, the TagProcessor will check if a tag is already present in the tag metadata table, and will not process the tag if it is.}
}}


\examples{
data_directory = here::here() # Root data directory

# Initialize the object
tp__ = TagProcessor(d = data_directory)

# Process the directory
tp__$process(con = con, overwrite = F)

# Build the processing report
report = tp__$build_report()
print(report)
}
